requirements.txt
pynacl>=1.5.0
pycryptodome>=3.19.0
